requirements-multimodal.txt
# v1.3.0 Multi-Modal Memory
pillow>=10.0.0
torch>=2.0.0
transformers>=4.30.0
openai-clip>=1.0.1
chromadb>=0.4.0
whisper>=1.1.10
opencv-python>=4.8.0
